create base object Crawler
